+++ activerecord/lib/active_record.rb
@@ -37,6 +37,7 @@ require "active_record/errors"
 module ActiveRecord
   extend ActiveSupport::Autoload
 
+  autoload :Alert
   autoload :Base
   autoload :Callbacks
   autoload :ConnectionHandling